add loading icon, fixed small glitches
