+++ src/components/Event-Scroll-Bar/EventScroll.jsx
@@ -2,10 +2,10 @@ import React from "react";
 
 import { default as button } from '../../assets/images/Events/button.png';
 
-import { EventScrollItem } from "../EventScrollItem/event-scroll-item.component";
-import './event-scroll.styles.scss';
+import { EventScrollItem } from "../EventScrollItem/EventScrollItem";
+import './EventScroll.scss';
 
-const EventScrollBar = ({ visible, loadless6, loadMore6, currIdx, events, loadPage }) => {
+const EventScrollBar = ({ visible, loadless6, loadMore6, currIdx, events, loadPage, isSelected}) => {
 
     return (
         <div className="scroll-bar-button">
@@ -17,7 +17,7 @@ const EventScrollBar = ({ visible, loadless6, loadMore6, currIdx, events, loadPa
             <div className="scroll-bar">
                 {
                     events.slice(visible, visible+7).map((event,idx) => (
-                        <EventScrollItem event = {event} idx = {idx} currIdx = {currIdx} visible={visible} loadPage={loadPage} />
+                        <EventScrollItem key={events[idx].id} isSelected = {isSelected} event = {event} idx = {idx} currIdx = {currIdx} visible={visible} loadPage={loadPage} />
                     ))
                 }
             </div>
