+++ src/components/EventScrollItem/EventScrollItem.scss
