+++ src/containers/Events/Events.scss
@@ -1,4 +1,6 @@
 .event-outer {
+    box-sizing: border-box;
+    padding: 4% 0;
     user-select: none;
     row-gap: 4%;
     height: 100vh;
@@ -18,5 +20,20 @@
         width: 55%;
         display: flex;
         justify-content: space-between;
+    }
+}
+
+.loading-logo{
+    height: 120px;
+    width: 120px;
+    animation: loading 2s infinite linear;
+}
+
+@keyframes loading {
+    from{
+        transform: rotate(0deg);
+    }
+    to{
+        transform: rotate(359deg);
     }
 }
